move bulk actions to saparate component
